refactor: names and folders
