+++ packages/contracts/test/10_unit-testing/base/11_majority-voting.ts
